+++ plume/src/test/kotlin/io/github/plume/oss/drivers/JanusGraphDriverIntTest.kt
@@ -31,16 +31,16 @@ import io.github.plume.oss.TestDomainResources.Companion.typeParameterVertex
 import io.github.plume.oss.TestDomainResources.Companion.typeRefVertex
 import io.github.plume.oss.TestDomainResources.Companion.unknownVertex
 import io.github.plume.oss.domain.exceptions.PlumeSchemaViolationException
-import io.github.plume.oss.graphio.GraphMLWriter
 import io.github.plume.oss.util.SootToPlumeUtil
 import io.shiftleft.codepropertygraph.generated.EdgeTypes.*
+import io.shiftleft.codepropertygraph.generated.NodeKeyNames.NAME
+import io.shiftleft.codepropertygraph.generated.NodeTypes.FILE
 import io.shiftleft.codepropertygraph.generated.nodes.*
 import org.apache.logging.log4j.LogManager
 import org.junit.jupiter.api.*
 import org.junit.jupiter.api.Assertions.*
 import overflowdb.Graph
 import scala.Option
-import java.io.FileWriter
 import kotlin.properties.Delegates
 
 class JanusGraphDriverIntTest {
@@ -202,6 +202,17 @@ class JanusGraphDriverIntTest {
             driver.addVertex(v2)
             assertTrue(driver.exists(v1))
             assertTrue(driver.exists(v2))
+        }
+
+        @Test
+        fun updateVertexTest() {
+            driver.addVertex(fileVertex)
+            assertTrue(driver.exists(fileVertex))
+            driver.getWholeGraph()
+                .use { g -> assertTrue(g.nodes(fileVertex.id()).asSequence().any { it.property(NAME) == STRING_1 }) }
+            driver.updateVertexProperty(fileVertex.id(), FILE, NAME, STRING_2)
+            driver.getWholeGraph()
+                .use { g -> assertTrue(g.nodes(fileVertex.id()).asSequence().any { it.property(NAME) == STRING_2 }) }
         }
     }