+++ plume/src/test/kotlin/io/github/plume/oss/util/DiffGraphUtilTest.kt
@@ -2,6 +2,7 @@ package io.github.plume.oss.util
 
 import io.github.plume.oss.TestDomainResources
 import io.github.plume.oss.TestDomainResources.Companion.STRING_1
+import io.github.plume.oss.TestDomainResources.Companion.STRING_2
 import io.github.plume.oss.TestDomainResources.Companion.fileVertex
 import io.github.plume.oss.TestDomainResources.Companion.methodVertex
 import io.github.plume.oss.drivers.DriverFactory
@@ -10,11 +11,13 @@ import io.github.plume.oss.drivers.TinkerGraphDriver
 import io.shiftleft.codepropertygraph.generated.EdgeTypes.SOURCE_FILE
 import io.shiftleft.codepropertygraph.generated.NodeKeyNames.NAME
 import io.shiftleft.codepropertygraph.generated.nodes.File
+import io.shiftleft.codepropertygraph.generated.nodes.StoredNode
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import overflowdb.Edge
+import overflowdb.Node
 import scala.Tuple2
 import scala.collection.immutable.Seq
 import scala.jdk.CollectionConverters
@@ -86,6 +89,22 @@ class DiffGraphUtilTest {
 
         DiffGraphUtil.processDiffGraph(driver, builder.build())
         driver.getWholeGraph().use { g -> assertFalse(g.edges().hasNext()) }
+    }
+
+    @Test
+    fun updateNodePropertyTest() {
+        driver.addVertex(fileVertex)
+        driver.getWholeGraph()
+            .use { g -> assertTrue(g.nodes(fileVertex.id()).asSequence().any { it.property(NAME) == STRING_1 }) }
+
+        val builder = io.shiftleft.passes.DiffGraph.newBuilder()
+        val storedNode: StoredNode
+        driver.getWholeGraph().use { g -> storedNode = g.node(fileVertex.id()) as StoredNode }
+        builder.addNodeProperty(storedNode, NAME, STRING_2)
+
+        DiffGraphUtil.processDiffGraph(driver, builder.build())
+        driver.getWholeGraph()
+            .use { g -> assertTrue(g.nodes(fileVertex.id()).asSequence().any { it.property(NAME) == STRING_2 }) }
     }
 
 }